services/gateway-service/requirements.txt
fastapi
uvicorn[standard]
websockets
pika
requests
pyjwt
httpx